Add spice_auto.html for automated console display
Add a spice_auto.html which removes the form inputs and obtains
connection details from query parameters. The optional 'token'
parameter is an opaque string passed to the websockets server
to allow it to auto-determine the correct console. Otherwise
the host/port/password/etc can be provided explicitly.

This is sufficient to integrate SPICE HTML5 with OpenStack

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>