Provide two levels of workaround for IE 10 cursor support.
The first approach is to use a moving image to simulate the cursor.
This does not work well, so a second approach, involving static
cursors is provided.  That requires a site administrator to precreate
all the cursors that can be used by a particular application set.

A warning message strongly suggests using a different browser as well.